+++ js/data.js
@@ -1,0 +1,28 @@
+const DESCRIPTIONS = [
+  'Хорошая фотография',
+  'Моя фотография',
+  'Отличная фотография',
+  'Великолепная фотография!',
+  'Замечательная фотография',
+  'Слишком короткое описание',
+  'Слишком большое описание',
+  'Тут что-то написано',
+  'Ой, как много текста!',
+  'Нечего и добавить',
+];
+
+const MESSAGES = [
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+];
+
+const NAMES = [
+  'Алексей',
+  'Елизавета',
+  'Максим',
+  'Илья',
+  'Ирина',
+  'Анна',
+];
+
+export {DESCRIPTIONS, MESSAGES, NAMES};
